Fixed my convenience function which was moved to nif and added nifnized gen_rand_float
Two different things in the same commit, sorry.
Added the possibility to send only state to gen_rand32/1 again,
this functionality was lost in rebase of dev.

Added the same functionality for gen_rand_float/1.
Actually this is slower than previously, at least in my measurements.
I.e. the list variant was faster.